Fix reading WAIT_TIME from env var as int, and add support for non-int-second input

+++ flask_boiler/testing_utils.py
@@ -9,7 +9,7 @@ try:
     import os
     wt = os.getenv("WAIT_TIME")
     if wt is not None:
-        DEFAULT_WAIT_TIME = wt
+        DEFAULT_WAIT_TIME = int(wt)
 except Exception as e:
     """
     TODO: find the best practice to handle a "black swan" error 
@@ -18,7 +18,7 @@ except Exception as e:
 
 
 def _wait(secs=DEFAULT_WAIT_TIME, factor=1):
-    secs = int(secs * factor)
+    secs = secs * factor
     time.sleep(secs)
     stats["total"] += secs
     logging.debug(msg=f"waited for {secs} seconds, "
